Rewrite /ifconfig
Change /ifconfig so that it uses the ipconfig binary.
If /bin/ipconfig exists and the IP address of the machine
is known, use ipconfig to configure the interface with
the given IP address, gateway and netmask.

If /bin/ipconfig exists and the IP address of the machine
is NOT known, then run ipconfig and try to get a DHCP
lease.

If /bin/ipconfig does NOT exist and the IP address is known,
then configure the interface manually.

If /bin/ipconfig does not exist and the IP address, use
a default configuration suitable for QEMU.

Use ipconfig for the loopback interface if it exists,
otherwise configure the loopback interface manually.

Also clean up the script in several ways:

1. Be consisent with quotes: use single quotes pretty
   uniformly throughout as opposed to a mixture of quotes
   and \-escaping #'s. Use double quotes when interpolating
   a variable into a string, of course.
2. Remove some old cruft and outdated comments.  It appeared
   that some of the comments dated from e.g. inferno or an
   earlier version of the system.
3. Move the 'default' configuration for network directly
   into /ifconfig.

Tested: Booting in various configurations.
Change-Id: I11aadb6f571c04c692d46a4794957245feb23d20

Signed-off-by: Dan Cross <crossd@gmail.com>