Modify 'ipconfig' to send an event to 'daemonize'.
As per the previous commit, modify `ipconfig` to participate
in the new `daemonize` protocol: use pthreads for the ipv6 router
advertisement/solitication and ipv6 DHCP loop maintenance threads.
Send an event to our parent when we exit or go into the background.

Change `ifconfig` to use daemonize and the new ipconfig facility.

Change-Id: I60735d0427c5ad580522cd7e5ac210f8d2d5ae4b
Signed-off-by: Dan Cross <crossd@gmail.com>